Added promotion service for automatic point increase code clean up

+++ CorruptOSBot/Program.cs
@@ -21,101 +21,69 @@ namespace CorruptOSBot
 {
     internal class Program
     {
-        private static void Main(string[] args)
-        {
-            // Call the Program constructor, followed by the
-            // MainAsync method and wait until it finishes (which should be never).
-            new Program().MainAsync().GetAwaiter().GetResult();
-        }
-
-        private readonly DiscordSocketClient _client;
+        private readonly DiscordSocketClient client;
 
         public static DateTime OnlineFrom { get; set; }
 
-        // Keep the CommandService and DI container around for use with commands.
-        // These two types require you install the Discord.Net.Commands package.
-        private readonly CommandService _commands;
+        private readonly CommandService commands;
+        private readonly IServiceProvider services;
+        private readonly Dictionary<ulong, Func<SocketMessage, IDiscordClient, Task>> channelInterceptors;
 
-        private readonly IServiceProvider _services;
-        private readonly Dictionary<ulong, Func<SocketMessage, Discord.IDiscordClient, Task>> _channelInterceptors;
-        private List<IService> _activeServices;
+        private List<IService> activeServices;
+
+        private static void Main(string[] args)
+        {
+            new Program().MainAsync().GetAwaiter().GetResult();
+        }
 
         private Program()
         {
             OnlineFrom = DateTime.Now;
 
-            _client = new DiscordSocketClient(new DiscordSocketConfig
+            client = new DiscordSocketClient(new DiscordSocketConfig
             {
-                // How much logging do you want to see?
                 LogLevel = LogSeverity.Info,
-
-                // If you or another service needs to do anything with messages
-                // (eg. checking Reactions, checking the content of edited/deleted messages),
-                // you must set the MessageCacheSize. You may adjust the number as needed.
-                //MessageCacheSize = 50,
-
-                // If your platform doesn't have native WebSockets,
-                // add Discord.Net.Providers.WS4Net from NuGet,
-                // add the `using` at the top, and uncomment this line:
-                //WebSocketProvider = WS4NetProvider.Instance
+                AlwaysDownloadUsers = true
             });
 
-            _commands = new CommandService(new CommandServiceConfig
+            commands = new CommandService(new CommandServiceConfig
             {
-                // Again, log level:
                 LogLevel = LogSeverity.Info,
-
-                // There's a few more properties you can set,
-                // for example, case-insensitive commands.
                 CaseSensitiveCommands = false,
             });
 
-            // Subscribe the logging handler to both the client and the CommandService.
-            _client.Log += Log;
-            _commands.Log += Log;
-
-            // Setup your DI container.
-            _services = ConfigureServices();
+            client.Log += Log;
+            commands.Log += Log;
+            services = ConfigureServices();
 
             ConfigHelper.Init();
 
-            _channelInterceptors = ConfigureChannelInterceptors();
-
-            _activeServices = ConfigureActiveServices();
+            channelInterceptors = ConfigureChannelInterceptors();
+            activeServices = ConfigureActiveServices();
         }
 
         private List<IService> ConfigureActiveServices()
         {
             var result = new List<IService>
             {
-                new PVMRoleService(_client),
-                new AchievementService(_client),
-                new TopKCService(_client),
-                new HeartbeatService(_client),
-                new SotWService(_client)
+                new PVMRoleService(client),
+                new AchievementService(client),
+                new TopKCService(client),
+                new HeartbeatService(client),
+                new SotWService(client),
+                new PromotionService(client)
             };
 
             return result;
         }
 
-        // If any services require the client, or the CommandService, or something else you keep on hand,
-        // pass them as parameters into this method as needed.
-        // If this method is getting pretty long, you can seperate it out into another file using partials.
         private static IServiceProvider ConfigureServices()
         {
             var map = new ServiceCollection();
-            // Repeat this for all the service classes
-            // and other dependencies that your commands might need.
-            //.AddSingleton(new SomeServiceClass());
-
-            // When all your required services are in the collection, build the container.
-            // Tip: There's an overload taking in a 'validateScopes' bool to make sure
-            // you haven't made any mistakes in your dependency graph.
+            
             return map.BuildServiceProvider();
         }
 
-        // Example of a logging handler. This can be re-used by addons
-        // that ask for a Func<LogMessage, Task>.
         public static Task Log(LogMessage message)
         {
             return LogHelper.Log(message);
@@ -123,36 +91,30 @@ namespace CorruptOSBot
 
         private async Task MainAsync()
         {
-            // Centralize the logic for commands into a separate method.
-
             await InitCommands();
 
             RootAdminManager.Init();
 
-            // Login and connect.
             if (!ConfigHelper.IsDebugMode)
             {
-                await _client.LoginAsync(TokenType.Bot, ConfigHelper.GetSettingProperty("DiscordToken"));
+                await client.LoginAsync(TokenType.Bot, ConfigHelper.GetSettingProperty("DiscordToken"));
             }
             else
             {
-                await _client.LoginAsync(TokenType.Bot, ConfigHelper.GetSettingProperty("TestDiscordToken"));
+                await client.LoginAsync(TokenType.Bot, ConfigHelper.GetSettingProperty("TestDiscordToken"));
             }
 
-            await _client.StartAsync();
-
-            //ReactionManager.Init();
+            await client.StartAsync();
 
             await LoadMemoryCache();
 
-            await StartServiceThreads(_client);
+            StartServiceThreads(client);
 
             foreach (var item in ToggleStateManager.GetToggleStates().OrderBy(X => X.Type).ThenBy(x => x.Functionality))
             {
                 await Log(new LogMessage(LogSeverity.Info, "Toggle states:", string.Format("[{2}] {0}: {1}", item.Functionality, item.Toggled, item.Type)));
             }
 
-            // Wait infinitely so your bot actually stays connected.
             await Task.Delay(Timeout.Infinite);
         }
 
@@ -165,12 +127,9 @@ namespace CorruptOSBot
             }
         }
 
-#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
-
-        private async Task StartServiceThreads(DiscordSocketClient client)
-#pragma warning restore CS1998 // Async method lacks 'await' operators and will run synchronously
+        private void StartServiceThreads(DiscordSocketClient client)
         {
-            foreach (var _activeService in _activeServices)
+            foreach (var _activeService in activeServices)
             {
                 new Thread(() =>
                 {
@@ -179,6 +138,7 @@ namespace CorruptOSBot
                         Thread.Sleep(_activeService.BeforeTriggerTimeInMS);
 
                         Thread.CurrentThread.IsBackground = true;
+
                         if (!ConfigHelper.IsDebugMode)
                         {
                             _activeService.Trigger(client);
@@ -190,18 +150,16 @@ namespace CorruptOSBot
             }
         }
 
-        private Dictionary<ulong, Func<SocketMessage, Discord.IDiscordClient, Task>> ConfigureChannelInterceptors()
+        private Dictionary<ulong, Func<SocketMessage, IDiscordClient, Task>> ConfigureChannelInterceptors()
         {
-            var result = new Dictionary<ulong, Func<SocketMessage, Discord.IDiscordClient, Task>>();
+            var result = new Dictionary<ulong, Func<SocketMessage, IDiscordClient, Task>>();
 
             AddToChannelInterceptorDictionary("suggestions", new SuggestionInterceptor().Trigger, result);
 
             return result;
         }
 
-        private void AddToChannelInterceptorDictionary(string channelName,
-            Func<SocketMessage, Discord.IDiscordClient, Task> func,
-            Dictionary<ulong, Func<SocketMessage, Discord.IDiscordClient, Task>> dictionaryToAddTo)
+        private void AddToChannelInterceptorDictionary(string channelName, Func<SocketMessage, IDiscordClient, Task> func, Dictionary<ulong, Func<SocketMessage, IDiscordClient, Task>> dictionaryToAddTo)
         {
             var channelId = ChannelHelper.GetChannelId(channelName);
             if (channelId != 0)
@@ -212,51 +170,37 @@ namespace CorruptOSBot
 
         private async Task InitCommands()
         {
-            // Either search the program and add all Module classes that can be found.
-            // Module classes MUST be marked 'public' or they will be ignored.
-            // You also need to pass your 'IServiceProvider' instance now,
-            // so make sure that's done before you get here.
-            //await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
+            await commands.AddModuleAsync<AdminModule>(services);
+            await commands.AddModuleAsync<AccountModule>(services);
+            await commands.AddModuleAsync<HelpModule>(services);
+            await commands.AddModuleAsync<BossKcModule>(services);
+            await commands.AddModuleAsync<KcModule>(services);
+            await commands.AddModuleAsync<RSNModule>(services);
+            await commands.AddModuleAsync<ScoreModule>(services);
+            await commands.AddModuleAsync<WoMModule>(services);
+            await commands.AddModuleAsync<PVMModule>(services);
+            await commands.AddModuleAsync<RoleModule>(services);
+            await commands.AddModuleAsync<CalendarModule>(services);
 
-            // Or add Modules manually if you prefer to be a little more explicit:
-            await _commands.AddModuleAsync<AdminModule>(_services);
-            await _commands.AddModuleAsync<AccountModule>(_services);
-            await _commands.AddModuleAsync<HelpModule>(_services);
-            await _commands.AddModuleAsync<BossKcModule>(_services);
-            await _commands.AddModuleAsync<KcModule>(_services);
-            await _commands.AddModuleAsync<RSNModule>(_services);
-            await _commands.AddModuleAsync<ScoreModule>(_services);
-            await _commands.AddModuleAsync<WoMModule>(_services);
-            await _commands.AddModuleAsync<PVMModule>(_services);
-            await _commands.AddModuleAsync<PromotionModule>(_services);
-            await _commands.AddModuleAsync<CalendarModule>(_services);
-
-            // Note that the first one is 'Modules' (plural) and the second is 'Module' (singular).
-
-            // Subscribe a handler to see if a message invokes a command.
-            _client.MessageReceived += HandleCommandAsync;
-            _client.UserJoined += _client_UserJoined;
-            _client.UserLeft += _client_UserLeft;
-            _client.UserBanned += _client_UserBanned;
-            _client.ReactionAdded += _client_ReactionAdded;
-            _client.CurrentUserUpdated += _client_CurrentUserUpdated;
+            client.MessageReceived += HandleCommandAsync;
+            client.UserJoined += Client_UserJoined;
+            client.UserLeft += Client_UserLeft;
+            client.UserBanned += Client_UserBanned;
+            client.ReactionAdded += Client_ReactionAdded;
+            client.CurrentUserUpdated += Client_CurrentUserUpdated;
         }
 
-        private async Task _client_CurrentUserUpdated(SocketSelfUser arg1, SocketSelfUser arg2)
+        private async Task Client_CurrentUserUpdated(SocketSelfUser arg1, SocketSelfUser arg2)
         {
             await Log(new LogMessage(LogSeverity.Info, "Users", "_client_CurrentUserUpdated"));
         }
 
-        private async Task _client_ReactionAdded(Cacheable<IUserMessage, ulong> arg1, ISocketMessageChannel arg2, SocketReaction arg3)
+        private async Task Client_ReactionAdded(Cacheable<IUserMessage, ulong> arg1, ISocketMessageChannel arg2, SocketReaction arg3)
         {
-            //Console.WriteLine(string.Format("Reaction: {0}: {1} [{2}]", arg3.Emote.Name, arg3.Emote.ToString(), ((Discord.Emote)arg3.Emote).Id));
-
             await Log(new LogMessage(LogSeverity.Info, "Users", string.Format("Reaction: {0}: {1}", arg3.Emote.Name, arg3.Emote.ToString())));
-
-            //await ReactionManager.ReactionPosted(_client, arg1, arg2, arg3);
         }
 
-        private async Task _client_UserJoined(SocketGuildUser arg)
+        private async Task Client_UserJoined(SocketGuildUser arg)
         {
             if (ToggleStateManager.GetToggleState(Constants.EventUserJoined))
             {
@@ -265,7 +209,7 @@ namespace CorruptOSBot
             }
         }
 
-        private async Task _client_UserLeft(SocketGuildUser arg)
+        private async Task Client_UserLeft(SocketGuildUser arg)
         {
             if (ToggleStateManager.GetToggleState(Constants.EventUserLeft))
             {
@@ -275,7 +219,7 @@ namespace CorruptOSBot
             }
         }
 
-        private async Task _client_UserBanned(SocketUser arg1, SocketGuild arg2)
+        private async Task Client_UserBanned(SocketUser arg1, SocketGuild arg2)
         {
             if (ToggleStateManager.GetToggleState(Constants.EventUserBanned))
             {
@@ -291,9 +235,7 @@ namespace CorruptOSBot
             if (msg == null) return;
 
             // We don't want the bot to respond to itself or other bots.
-            if (msg.Author.Id == _client.CurrentUser.Id || msg.Author.IsBot) return;
-
-            // block in debug mode except admin
+            if (msg.Author.Id == client.CurrentUser.Id || msg.Author.IsBot) return;
 
             await BlockMessageIfDebugMode(msg);
 
@@ -306,33 +248,20 @@ namespace CorruptOSBot
                 await Log(new LogMessage(LogSeverity.Error, "chatlog", e.Message, e));
             }
 
-            // check for channels, if its for a targeted channel, intercept
             int posX = 0;
-            if (_channelInterceptors.ContainsKey(arg.Channel.Id) && !msg.HasCharPrefix('!', ref posX))
+            if (channelInterceptors.ContainsKey(arg.Channel.Id) && !msg.HasCharPrefix('!', ref posX))
             {
-                await _channelInterceptors[arg.Channel.Id](arg, _client);
+                await channelInterceptors[arg.Channel.Id](arg, client);
             }
             else
             {
-                // Create a number to track where the prefix ends and the command begins
                 int pos = 0;
-                // Replace the '!' with whatever character
-                // you want to prefix your commands with.
-                // Uncomment the second half if you also want
-                // commands to be invoked by mentioning the bot instead.
-                if (msg.HasCharPrefix('!', ref pos) /* || msg.HasMentionPrefix(_client.CurrentUser, ref pos) */)
+
+                if (msg.HasCharPrefix('!', ref pos))
                 {
-                    // Create a Command Context.
-                    var context = new SocketCommandContext(_client, msg);
+                    var context = new SocketCommandContext(client, msg);
+                    var result = await commands.ExecuteAsync(context, pos, services);
 
-                    // Execute the command. (result does not indicate a return value,
-                    // rather an object stating if the command executed successfully).
-                    var result = await _commands.ExecuteAsync(context, pos, _services);
-
-                    // Uncomment the following lines if you want the bot
-                    // to send a message if it failed.
-                    // This does not catch errors from commands with 'RunMode.Async',
-                    // subscribe a handler for '_commands.CommandExecuted' to see those.
                     if (!result.IsSuccess && result.Error != CommandError.UnknownCommand)
                     {
                         await LogHelper.Log(new LogMessage(LogSeverity.Error, msg.Content, result.ErrorReason));
@@ -343,10 +272,11 @@ namespace CorruptOSBot
 
         private async Task BlockMessageIfDebugMode(SocketUserMessage msg)
         {
-            if (ConfigHelper.IsDebugMode)
+            var isDeveloper = msg.Author.Id == 108710294049542144 || msg.Author.Id == 391595176314798090; //GMKirby | Carn3asada Discord Id
+
+            if (ConfigHelper.IsDebugMode && !isDeveloper)
             {
-                //block!
-                var context = new SocketCommandContext(_client, msg);
+                var context = new SocketCommandContext(client, msg);
                 var message = await context.Channel.SendMessageAsync("The corrupt bot is currently in development mode.");
 
                 await msg.DeleteAsync();
